pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["viniciusmth <viniciusmatheus40@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.13.*"
uvicorn = "^0.34.0"
fastapi = {extras = ["standard"], version = "^0.115.6"}
sqlalchemy = "^2.0.36"
pydantic-settings = "^2.7.1"
alembic = "^1.14.0"

[tool.poetry.group.dev.dependencies]
gh = "^0.0.4"
ignr = "^2.2"
pipx = "^1.7.1"
ruff = "^0.8.4"
pytest = "^8.3.4"
pytest-cov = "^6.0.0"
taskipy = "^1.14.1"

[tool.pytest.ini_options]
pythonpath='.'
addopts='-p no:warnings'

[tool.taskipy.tasks]
run = 'fastapi dev backend/app.py'

pre_test = 'task lint'
test = 'pytest --cov=backend -vv'
post_test = 'coverage html'

lint = 'ruff check . && ruff check . --diff'
format = 'ruff check . --fix && ruff format .'

[tool.ruff]
line-length = 200
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview=true
quote-style='single'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
